nexus-ai-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymupdf==1.23.7
transformers==4.35.2
torch>=2.1.1
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2
pypdf==3.17.1
google-generativeai==0.8.4
langchain==0.1.0
pydantic==2.5.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
chromadb==0.4.18 
passlib[bcrypt]==1.7.4
PyJWT==2.9.0